PathFinding 2x2 correct movement with static obstacles

+++ Assets/Game/Scripts/AI/IPathFindingGraph.cs
@@ -2,7 +2,7 @@ using System.Collections.Generic;
 
 public interface IPathFindingGraph
 {
-    void GeneratePathTo(int x, int y);
+    void GeneratePathTo(float x, float y);
     void SetDynamicObstacleNodes();
     List<Node> GetDynamicObstacleNodes();
     void GeneratePathFindingGraph();
